Added conversion from buffer in Python to cv::Mat in C++, and Vec2f/Vec4f bindings
This allows buffer objects (e.g. arrays) to be converted to eos.Mat (and eos.VecXf), which will then subsequently be converted to cv::Mat/cv::VecXf in C++.
Added mostly support for 1-channel CV_32F and CV_64F types.
The explicit conversion that's required in Python is a bit cumbersome, but I can live with it for now.