Added viewport flipping to linear SOP estimation
Adjusted RenderingParameters constructor accordingly.

This is useful to get correct angle estimates when, as in OpenCV, the viewport is upside-down.
It could maybe also be solved by negating the 2nd and 3rd row of the rotation matrix in RenderingParameters, but I couldn't get the projection right. This is probably nicer.