Added .clang-format file
This is more or less the coding style that should be followed for new contributions.
I am happy if code differs from this style a bit if there is a valid reason for it.

It's a rough first version of this file, it may change a bit over time, if required.