Removed json-to-cereal-binary utility app
This was in the past used for converting a small BFM json model to a cereal binary. However, it was never satisfactory.
I think it's also used to import the blendshapes from Matlab into C++, so we'll need it again at some point. However at the moment it does not compile as we don't have a json serialize() (cereal) function for Eigen::Matrix.