Fixed storage order bug with Eigen in the shape and blendshape fitting
When switching the shape and blendshape fitting to use the new Eigen models, these matrices weren't correctly mapped to cv::Mat. Eigen uses col-major storage order by default, and OpenCV row-major - so we now convert to a row-major Eigen matrix before mapping to a cv::Mat.
This caused the fitting to produce faulty/meaningless results since the switch to Eigen.
The output of the linear shape and blendshape fitting functions are now again identical before and after the switch to Eigen. (after 50 iterations, the result is marginally different, most likely because of small rounding differences.)