Initial implementation of the landmarksweighting
For the shape and expression (NNLS) fitting.
The required update for the NNLS solver is coming in the next commit.

I'm not sure whether it's 100% working yet. It doesn't converge (runs forever, or even crashes, because it can't add variables to the active set of the NNLS), if the sdev is set too low.
I set max_iter=50 for now (which still can crash if the sdev is set too low).

The sdev can't be put too low, I think it kind of acts like an anti-regularisation.
I think I should implement weighting for the camera fitting too.