Added new (and correct) script to convert the BFM2009
Added a python script that converts the Basel Face Model (BFM) 2009 to the eos format.
It is tested and compared against the figures from their AVSS 2009 paper, and the shape and colour variations match.

Removed the old converter scripts and apps (they had a mistake in them if I saw it correctly).