Added script to convert the BFM2017
Added a python script that converts the new Basel Face Model (BFM) 2017 to the eos format.
It is tested and compared against the figures from their AVSS 2009 paper, and the shape and colour variations match.

It does not convert the BFM2017 expression model, as eos does not yet support PCA expression models.