Changed MorphableModel so that it just works for the BFM2017
Removed <optional> and <variant> and all logic regarding it.
This results in the MorphableModel class now just being able to represent the BFM2017 with its PCA shape and expression models).
The BFM2017 needs to be converted with the Python script using this eos branch.
This is a first test to get the BFM2017 work on Xcode.