Remove g++-8 from allowed travis failures, add Xcode 9.2 to allowed failures
Probably some minor updated of gcc-8 fixed something and made it work (I think it has been working for a while now).
On the other hand, Apple frequently changes some stuff with Xcode and that now fails. It's probably not too hard to fix but I can't look into that right now.