Added texture triangulation member to MorphableModel, bumped class version to 4
To make it possible to have uv maps with seams, we also need to store the triangulation of the uv mapping's triangles, alongside the uv coordinates.
This commit adds the member variable `texture_triangle_indices` and bumps the MorphableModel CEREAL_CLASS_VERSION to 4. All previous models can still be read.
Also added texture_triangle_indices to the two constructors, and to the Python bindings.