Pass along texture_triangle_indices to sample_to_mesh in all MorphableModel member functions
Previously, get_mean and the draw_sample functions (except for the one that takes shape, colour and expression coefficients) would have returned a mesh without the texture triangle indices (if the MorphableModel contained them).
This might actually have been causing wrong code to be run or some bad memory access to happen in Release and RelWithDebInfo builds, because the assert that checks for that in sample_to_mesh(...) wouldn't have been triggered.