Changed strides in Image python bindings to account for row-major Image class
It seems from the comment in the code that with the introduction of the core::Image class in July 2018, I just copied the cv::Mat bindings file, and did only minor adjustments (and haven't adjusted the strides). The core::Image class stores the data in row-major order, so the strides would've likely changed. Indeed the `core::Image4u` image of `extract_texture(...)` was returned to Python transposed. And when I am testing the new `draw_wireframe(...)` returning an `Image3u`, it's all interleaved/wrong.
This change now fixes both and makes it so that for fitting on the example image, it works correctly.