Added project(...) function to PcaModel
This is to project a given data instance (e.g. a given shape instance) onto the PCA space, to get the resulting PCA coefficients.
Also added Python bindings for it.

There is one peculiarity that we seem to need to divide the resulting coefficients with their corresponding eigenvalues. I think that this shouldn't be needed, as we project using the rescaled PCA basis. This will need some further checking.