Removed eos::render::extract_texture(...)
This function only supported affine cameras, and it was time for it to be removed. The `v2::extract_texture` will move to the main namespace and be made the default.
I also noticed with this old function (the one being removed) that even though it computes a z-buffer, on a side image, with a 3DMM with ears, part of the occluding ear was being rendered.
The `TextureInterpolation` option is also removed, but `NearestNeighbour` was anyway the only one ever really tested and it was the default. I also think the view-angle weighting also only worked with `NearestNeighbour`.