Moved v2::extract_texture to the eos::render namespace
The function is mature enough to be the new default. It supports perspective cameras, as well as orthographic cameras. It uses the "v2" `Rasterizer` architecture, which is a much more general method than warping pixels using an affine warp, like the previous method did.

The new function has a slightly different API, in that a proper `view_model_matrix` and `projection_matrix` can be passed now.
It does **not** yet support outputting the view-angle in the alpha channel - this will need to be added.
It might also be slower than the old implementation, as it is not parallelised.