Support meshes without vertex colours in SoftwareRenderer
Previously, if a Mesh didn't contain `Mesh::colors`, it would try to index it nonetheless and create an out-of-bounds access.
This adds a get_color_or_zero(...) lambda that returns the color, if present, or an empty vector otherwise.

The v1 renderer got around this with an `if` that sets the vertex colour to (0.5, 0.5, 0.5) if it is not given, when building the list of vertices to render: https://github.com/patrikhuber/eos/blob/c9738098ee0ea57d7e3a7d4da7d7aaa933648e8f/include/eos/render/render.hpp#L161-L167.