Removed v1 renderer and added convenience function for v2 renderer
This removes the v1 renderer code and the detail/render_detail.hpp file, which is no longer needed.
Added two convenience functions to render.hpp so the v2 renderer (now in the default namespace) can be used in a very similar way to the v1 renderer.

The core of the v2 renderer is the same, but the architecture has been changed to mirror more closely the hardware (or OpenGL) pipeline, with vertex- and fragment shaders. Results should be exactly the same, and they have the same parameters.