Added a function to compute Euler (Tait-Bryan) angles from a rotation matrix using Eigen
We can use Eigen::Matrix3::eulerAngles(), but it uses a fairly unusual convention where the middle axis is constrained to [0, pi] (I think). This results in yaw angles being "explained" by pitch. There is an equivalent solution where the middle (pitch) axis is constrained to [-pi/2, pi/2], which is the solution we want. The formula in the code does that swapping (see the linked Eigen issue for more details).