Added build matrix to cmake.yml
Had to add an extra variable for the vcpkg cache path, since it is different on Windows. Ideally we'd probably use `$env:LOCALAPPDATA/vcpkg/archives/`, but I couldn't make it work, it would just display "$env:..." as a string.
Alternatively, we could set `%VCPKG_DEFAULT_BINARY_CACHE%` on Windows and `$VCPKG_DEFAULT_BINARY_CACHE` on Linux/macOS, and vcpkg will then use that path.

We use `fail-fast: false` so that the builds continue, if one of the builds fail (the default is "true").