Change vcpkg cache key from runner.OS to matrix.os
I'm not sure if vcpkg's binaries contain information about the compiler / platform version they've been built with, and I think it may mess up things if we mix the cached dependencies like this. The builds of the older OS'es also outputted an error at the end that they couldn't access the cache as it is already being accessed.

This will separate the caches, which should be a good thing to do in any case.