Include OpenCV imgcodecs instead of highgui in examples
It looks like cv::imwrite was in highgui in OpenCV2, thus we included that header. But for a while now, it's been in imgcodecs. Thus if OpenCV 3 or 4 is found, we were only looking for (and linking to) imgcodecs, but not highgui - but we were still including the highgui header, and this leads to a compiler error in most recent builds.