Set PYTHON_IS_DEBUG for CMAKE_BUILD_TYPE Debug
This fixes the issue that in debug builds, the linker would try and link to pythonxyz_d.lib, but complain that pythonxyz.lib was not found. It looks like setting PYTHON_IS_DEBUG is needed since Python 3.8.

The CMake if(...) probably won't work well for multi-configuration generators (like the Visual Studio Generator), but there's not much we can do (generator expressions don't work here).