cibuildwheel: Skip arm64 and universal wheels on cp37 and cp38
When building `cp38-macosx_universal2`, cibuildwheel fails with a delocate-error "Failed to find any binary with the required architecture: 'arm64'". This is probably because the built universal2 wheel for CPython 3.8 doesn't actually contain any arm64 code. It looks like on macOS, there isn't an official native CPython 3.8 installer for arm64 — so when we try to build a universal2 wheel for CPython 3.8 on an x86_64 runner, the build ends up producing an x86_64‑only wheel (even though its name suggests universal2). Then, when delocate‑wheel is run with the option to require both x86_64 and arm64, it can't find any arm64 binary, and it fails with that error.